MaxEntNucleotideBiases

Documentation for MaxEntNucleotideBiases.

MaxEntNucleotideBiases.ComputeEntropy

MaxEntNucleotideBiases.ComputeLoglikelihood

MaxEntNucleotideBiases.ComputeMinusEnergy

MaxEntNucleotideBiases.ComputeMinusEnergyLmodel3

MaxEntNucleotideBiases.ComputePressure

MaxEntNucleotideBiases.ComputeSymmetrizedKL

MaxEntNucleotideBiases.EvalLogZ

MaxEntNucleotideBiases.EvalLogZFast

MaxEntNucleotideBiases.GaugeMaskVariables

MaxEntNucleotideBiases.GenerateTM

MaxEntNucleotideBiases.GenerateTMLast

MaxEntNucleotideBiases.MetropolisSampling

MaxEntNucleotideBiases.ModelFit

MaxEntNucleotideBiases.ZerosumGauge

MaxEntNucleotideBiases._ComputeNObs

MaxEntNucleotideBiases._MCStep!

MaxEntNucleotideBiases._PreprocessSeqs

MaxEntNucleotideBiases.ComputeEntropy — Method

ComputeEntropy(L::Int, model_pars::Dict{String, Float64}; fast::Bool=false)

Compute the entropy of the model with parameters model_pars and having length L. The computation exploit the derivatives of the partition function, that can be approximated to make it faster (using fast=true), although it should be very quick in any case.

MaxEntNucleotideBiases.ComputeLoglikelihood — Method

ComputeLoglikelihood(seq::String, model_pars::Dict{String, Float64}; logZ=missing)

Given a sequence seq and the model parameters model_pars, compute the log-likelihood (energy minus log of Z) of this sequence. logZ can be passed directly if pre-computed, otherwise is it computed each time this function is called.

MaxEntNucleotideBiases.ComputeMinusEnergy — Method

ComputeMinusEnergy(seq::String, model_pars::Dict{String, Float64})

Given a sequence seq and the model parameters model_pars, compute (minus) the energy of this sequence.

MaxEntNucleotideBiases.ComputeMinusEnergyLmodel3 — Method

ComputeMinusEnergyLmodel3(seq::String, model_pars::Dict{String, Float64})

Given a sequence seq and the model parameters model_pars, compute (minus) the energy of this sequence. It only works for motifs of length 3, and it is slightly faster than ComputeMinusEnergy.

MaxEntNucleotideBiases.ComputePressure — Method

ComputePressure(L::Int, model_pars::Dict{String, Float64}; fast::Bool=false)

The pressure on a genome can be quantified as how the genome is different from a random uniform one. Therefore, this function compute the difference in the entropy of a random uniform model and the model described with model_pars, with lenght L, rescaled with the entropy of the uniform model. Setting fast=true allows for a quicker estimation of the entropy of the model (it can be useful for extremely large sequences).

MaxEntNucleotideBiases.ComputeSymmetrizedKL — Method

ComputeSymmetrizedKL(L::Int, model1_pars::Dict{String, Float64}, 
                       model2_pars::Dict{String, Float64}; fast::Bool=false)

Compute the symmetrized version of the Kullback-Leibler divergence between the distributions defined by model1_pars and model2_pars. Setting fast=true allows for a quicker estimation of the KL divergence (it can be useful for extremely large sequences).

MaxEntNucleotideBiases.EvalLogZ — Method

EvalLogZ(model_pars::Dict{String, Float64}, L::Int)

Compute the partition function of a model of length L through the transfer matrix method. model_pars is a dict of motif => parameter in the Hamiltonian.

MaxEntNucleotideBiases.EvalLogZFast — Method

EvalLogZFast(model_pars::Dict{String, Float64}, L::Int)

Compute the partition function of a model of length L by taking the real part of the largest eigenvalue of the transfer matrix. model_pars is a dict of motif => parameter in the Hamiltonian.

MaxEntNucleotideBiases.GaugeMaskVariables — Method

GaugeMaskVariables(motifs::Vector{String})

This function returns a Array{Bool} corresponding to motifs that have to be inferred after all motifs containing Ts are fixed to 0 thanks to gauge transformations.

MaxEntNucleotideBiases.GenerateTM — Method

GenerateTM(model_pars::Dict{String, Float64})

Return the trasfer matrix used for the computation of the partition function. The first index correspond to the left-most nucleotide, the last index to the right-most one.

MaxEntNucleotideBiases.GenerateTMLast — Method

GenerateTMLast(model_pars::Dict{String, Float64})

Return the last transfer matrix used for the computation of the partition function. The first index correspond to the left-most nucleotide, the last index to the right-most one.

MaxEntNucleotideBiases.MetropolisSampling — Method

MetropolisSampling(L::Int, model_pars::Dict{String, Float64}; 
                    beta::Float64=1.0, Nsamples::Int=1, Nsteps::Int=1, Ntherm::Int=L*10, 
                    startseq=missing)

Sample Nsample sequences of length L at inverse temperature beta using a model defined by the model parameters given in model_pars (dict of motifs => floats). Ntherm is the number of MonteCarlo steps done before starting collecting sequences, startseq is the starting sequence.

MaxEntNucleotideBiases.ModelFit — Function

ModelFit(seqs::Vector{String}, Lmotifs::Int, Lmodel::Union{Int,Missing}=missing; 
         add_pseudocount::Bool=false, tolerance::Float64=0.01, max_iter::Int=100, 
         verbose::Bool=true)
ModelFit(seq::String, Lmotifs::Int, Lmodel::Union{Int,Missing}=missing; 
         add_pseudocount::Bool=false, tolerance::Float64=0.01, max_iter::Int=100, 
         verbose::Bool=true)

Fit the model parameters, which are:

only 1-point functions (fields) if Lmotifs==1;

1-point and 2-point functions (fields, 2-mer forces) if Lmotifs==2;

1-point, 2-point and 3-point functions (fields, 2-mer and 3-mer forces) if Lmotifs==3.

Lmodel is the number of nucleotides used for the inference. If seqs is made of sequences of constant length (or if it is String and not a vector), the default option will use the sequences length as Lmodel; otherwise, the default behaviour is to take Lmodel = 5000.

If add_pseudocount, a single pseudocount is added for each observed number of nucleotides and dinucleotides.

If fast, the partition function is estimated through the top eigenvalue of the transfer matrix alone (much faster, but slightly less precise, expecially for short sequences).

Finally, tolerance and max_iter are parameters for the Newton-Raphson algorithm used to solve the system of equations.

MaxEntNucleotideBiases.ZerosumGauge — Method

ZerosumGauge(model_pars::Dict{String, Float64})

This function takes as input a dictionary describing the parmeters of a model and changes the gauge so that one and two point parameters are "zero-sum". It does not modify three point parameters.

MaxEntNucleotideBiases._ComputeNObs — Method

_ComputeNObs(seqs::Vector{String}, independent_motifs::Vector{String}, L::Int)

For each motif in independent_motifs, compute the number of observed motifs in each sequence in seqs, then divide by the sequence length, take the average of these intensive fractions over the sequences, and multiply by the model length L.

MaxEntNucleotideBiases._MCStep! — Method

_MCStep!(L::Int, mlk::Int, curr_seq::Vector{String}, 
          model_pars::Dict{String, Float64}, 
          beta::Float64)

Performs a MonteCarlo step; L is currseq length, mlk is the maximum length of the motifs considered, currseq is the starting sequence, model_pars are the model parameters (dict of motifs => floats), and beta is the inverse temperature.

MaxEntNucleotideBiases._PreprocessSeqs — Method

_PreprocessSeqs(seqs_in::Vector{String})

Pre-process the vector of sequences seqs_in, so that sequences only contains letters A, C, G, T. The pre-processed sequence vector is the returned.