MaxEntNucleotideBiases

Documentation for MaxEntNucleotideBiases.

Installation

This package is not registered. Install with:

import Pkg
Pkg.add(url="https://github.com/adigioacchino/MaxEntNucleotideBiases.jl")

Introduction and references

TBD

Examples

Basic usage

We will take sub-sequence from the Influenza H5N1 PB2 segment (strain used: A/Anhui/1/2005) and we will use it as working example to test the function of the package. We start by loading the package and the sequence:

using MaxEntNucleotideBiases
example_seq = "ATGGAGAGAATAAAAGAATTAAGGGATCTAATGTCACAGTCCCGCACTCGCGAGATACTAACAAAAACCACTGTGGACCATATGGCCATAATCAAGAAGTACACATCAGGAAGACAAGAGAAGAACCCTGCTCTCAGAATGAAATGGATGATGGCAATGAAATATCCAATCACAGCGGACAAGAGAATAACAGAGATGATTCCTGAAAGGAATGAACAAGGGCAGACGCTCTGGAGCAAGACAAATGATGCCGGATCGGACAGGTTGATGGTGTCTCCCTTAGCTGTAACTTGGTGGAATAGGAATGGGCCGACGACAAGTGCAGTCCATTATCCAAAGGTTTACAAAACATACTTTGAGAAGGCT"

The package allows to fit up to 1-, 2- or 3-mers (i.e. single nucleotide, dinucleotide or trinucleotide) forces. To do so, we will use the fitmodel function, whose second argument is the 'k' of the maximum k-mers we want to consider (1, 2 or 3).

We will start by fitting the single nucleotide biases:

julia> mod1 = MaxEntNucleotideBiases.fitmodel(example_seq, 1)MaxEntNucleotideBiases.NucleotideModel(["A", "C", "G", "T"], [-0.9973585549629368, -1.6541380913520074, -1.3917738268845161, -1.6541380913520074], 1)

The result is a NucleotideModel object, which contains the motifs, the fitted parameters and the value of 'k'. Similarly, we can fit dinucleotide and trinucleotide biases:

julia> mod2 = MaxEntNucleotideBiases.fitmodel(example_seq, 2)MaxEntNucleotideBiases.NucleotideModel(["A", "C", "G", "AA", "AC", "AG", "CA", "CC", "CG", "GA", "GC", "GG", "T", "AT", "CT", "GT", "TA", "TC", "TG", "TT"], [0.47810213581717065, -0.19372984275095523, -0.03740291016895336, -0.0707982808908727, -0.13444053340491877, -0.03225979088728337, 0.2004348089198602, 0.11573908379605727, -0.5805713126468183, 0.27971246720385734, -0.18635026657710094, 0.16714441678698094, -0.2469693828972621, 0.23749860518307497, 0.2643974199309007, -0.2605066174137374, -0.40934899523284474, 0.20505171618596243, 0.44568668674712064, -0.2413894077002383], 2)
julia> mod3 = MaxEntNucleotideBiases.fitmodel(example_seq, 3)MaxEntNucleotideBiases.NucleotideModel(["A", "C", "G", "AA", "AC", "AG", "CA", "CC", "CG", "GA"  …  "TCG", "TCT", "TGA", "TGC", "TGG", "TGT", "TTA", "TTC", "TTG", "TTT"], [0.572913248492579, -0.1792159045763034, -0.09335938885626413, -0.018687725512444797, -0.3002901265164418, 0.009228434394642937, 0.2322363379889537, 0.23385794534202153, -1.086820558445575, 0.3556933709895902  …  -0.12422685254954782, -0.036205853127477886, -0.43403484942914083, -0.6059729363750332, 0.17223128631340057, 0.8677764994907737, 0.7345198622907295, -0.8888703092329991, -0.18836551451264016, 0.3427159614549099], 3)

From a model we can obtain a dictionary (motif => force) with the get_forces_dict function:

julia> MaxEntNucleotideBiases.get_forces_dict(mod2)Dict{String, Float64} with 20 entries:
  "AT" => 0.237499
  "A"  => 0.478102
  "CA" => 0.200435
  "C"  => -0.19373
  "T"  => -0.246969
  "TG" => 0.445687
  "CC" => 0.115739
  "G"  => -0.0374029
  "GT" => -0.260507
  "TA" => -0.409349
  "GA" => 0.279712
  "TT" => -0.241389
  "AC" => -0.134441
  "CT" => 0.264397
  "GC" => -0.18635
  "AA" => -0.0707983
  "AG" => -0.0322598
  "GG" => 0.167144
  "CG" => -0.580571
  "TC" => 0.205052

Finally, we can compute several interesting quantities:

the energy of any sequence with a given model, with the compute_minus_energy function:

julia> L = 1000;
julia> seq = join(rand(["A","C","G","T"], L));
julia> MaxEntNucleotideBiases.compute_minus_energy(seq, mod3)-13.616383985405314

the log-likelihood of any sequence with a given model, with the compute_loglikelihood function:

julia> L = 1000;
julia> seq = join(rand(["A","C","G","T"], L));
julia> MaxEntNucleotideBiases.compute_loglikelihood(seq, mod2)-1463.3691585354652

Gauge choices

There are several gauge choices that can be used to fit the model. The default choice is the 'zero-sum' gauge, that is described in the paper. The other notable choice is the 'lattice gas gauge', which is again described in the paper and is implemented in this package. In particular, the results of the inference can be directly returned in the lattice gas gauge with the ZS_gauge argument passed as false to the fitmodel function:

julia> mod2 = MaxEntNucleotideBiases.fitmodel(example_seq, 2, ZS_gauge=false)MaxEntNucleotideBiases.NucleotideModel(["A", "C", "G", "AA", "AC", "AG", "CA", "CC", "CG", "GA", "GC", "GG", "T", "AT", "CT", "GT", "TA", "TC", "TG", "TT"], [1.03599994406514, 1.0054674916636468, 0.8775253574621688, -0.14033729854134125, -0.8183802624741945, -0.9568344905177173, 0.10399697652156595, -0.5950994600210442, -1.532044827025078, 0.7081786721502011, -0.3722847730495643, -0.25942506024664064, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], 2)

Moreover, the zero-sum gauge can be restored after the inference with the gauge_zerosum function:

julia> MaxEntNucleotideBiases.gauge_zerosum(mod2)MaxEntNucleotideBiases.NucleotideModel(["A", "C", "G", "AA", "AC", "AG", "CA", "CC", "CG", "GA", "GC", "GG", "T", "AT", "CT", "GT", "TA", "TC", "TG", "TT"], [0.47810213581717065, -0.19372984275095523, -0.03740291016895336, -0.0707982808908727, -0.13444053340491877, -0.03225979088728337, 0.2004348089198602, 0.11573908379605727, -0.5805713126468183, 0.27971246720385734, -0.18635026657710094, 0.16714441678698094, -0.2469693828972621, 0.23749860518307497, 0.2643974199309007, -0.2605066174137374, -0.40934899523284474, 0.20505171618596243, 0.44568668674712064, -0.2413894077002383], 2)

Entropy and related quantities

The simple structure of the models fitted with this package allows to compute exactly several quantities of interest that, in most other cases, require (uncontrolled) approximations. The first example is the entropy of the probability distribution defined with the model. It can be computed with the compute_entropy function, and since this quantity depends on the length of the sequences considered, a reference length must be passed as argument:

julia> L = 1000;
julia> MaxEntNucleotideBiases.compute_entropy(L, mod3)1274.7420936767367

The entropy of the model can be used to quantify how different the probability distribution defined by the model is from the uniform distribution, which in turn is a measure of the 'pressure' that is exerted on the nucleotides to be used in a certain way. In this package, the pressure ranges from 0 (unconstrained model) to 1 (fully constrained model), and it is computed with the compute_pressure function:

julia> L = 1000;
julia> MaxEntNucleotideBiases.compute_pressure(L, mod3)0.08046795152007868

Finally, it can be interesting to quantify how much two models are different from each other. This package address this problem with the compute_symmetrized_kl function, which computes the symmetrized Kullback-Leibler divergence between two models:

julia> L = 1000;
julia> mod3_unif = MaxEntNucleotideBiases.fitmodel(join(rand(["A","C","G","T"], L)), 3)MaxEntNucleotideBiases.NucleotideModel(["A", "C", "G", "AA", "AC", "AG", "CA", "CC", "CG", "GA"  …  "TCG", "TCT", "TGA", "TGC", "TGG", "TGT", "TTA", "TTC", "TTG", "TTT"], [-0.017002192810154526, -0.019791398869569043, 0.023156177213518536, -0.00032886983139374104, -0.017636061400274405, 0.13435888647393573, 0.09094347807494435, -0.04069950705142, -0.14556155437306534, -0.16040631935647226  …  0.2117039750006569, -0.19539319643368525, -0.13748594944487627, -0.06481228861657928, 0.07363972313435327, 0.12865851492710226, 0.08977484603323305, -0.0814377375994561, -0.2640402639987774, 0.2557031555650004], 3)
julia> MaxEntNucleotideBiases.compute_symmetrized_kl(L, mod3, mod3_unif)160.10994094977457

Sampling

Another interseting feature of this package is the possibility to sample sequences from an inferred model. This package implements a simple Metropolis routine to sample sequences from a model:

julia> L = 1000;
julia> MaxEntNucleotideBiases.sample_metropolis(L, mod3,
                                                Nsamples=10, Nsteps=3000, Ntherm=10_000, beta=1.)10-element Vector{String}:
 "AAATCCTTGGGCTGAAGCAATGAAGATTTAA" ⋯ 939 bytes ⋯ "TACATGAAGATGAGCCTTACTGGATGAGAA"
 "GGTCTCACAGACACAGGCACAGTGATTTTGG" ⋯ 939 bytes ⋯ "CAATGCATTATGAGACAGTAAGATCTACTC"
 "GCCTTACAGAACTTAATGGATAAGATCCAGA" ⋯ 939 bytes ⋯ "GAGCCTCATTAGCTTACAAGAGACAACATC"
 "GGCGGACAGGGCTGAGAGGCTGATTGAAACA" ⋯ 939 bytes ⋯ "TAGGAATATTGAAACCAACAAAAGAATATC"
 "GAGAGAAGCGCCGCGAAGTTTTAATCGACGG" ⋯ 939 bytes ⋯ "ATAAGAGCTGGGAATATACATATGTTTTAA"
 "TCAAAGAGATGAGTTAATCCCGCTGAGCTTA" ⋯ 939 bytes ⋯ "ATGTCCACAATGGAGAATACAAGAAAACAA"
 "GCCTCTAGATGATGGTCCATGGCGCTCCACA" ⋯ 939 bytes ⋯ "ATGGCCAGAATGGAAGAGACAAGAAGACAA"
 "GCTTATCAAACAATATCCAGATATCCCCACT" ⋯ 939 bytes ⋯ "TAAGTCAATTTGGACAGGAACAGGACACTG"
 "ATGAAAAAGGAACTTCAAAAATGAGGTCGAT" ⋯ 939 bytes ⋯ "AAACTAAGATCCGAAGAGAACCGGATGATG"
 "CACAGAATGATACTTTGACAACAACTATGAT" ⋯ 939 bytes ⋯ "ACTTACAGACTCCAAAACCCTTAGAGAAGG"

Model I/O

The model inference can take some time, so it can be useful to save the model on disk and load it later. This can be done with the save_model and load_model functions:

MaxEntNucleotideBiases.writemodel("mod3.menb", mod3)
mod3 = MaxEntNucleotideBiases.readmodel("mod3.menb")

API

MaxEntNucleotideBiases._compute_freqs

MaxEntNucleotideBiases._mcstep!

MaxEntNucleotideBiases._preprocess_seqs

MaxEntNucleotideBiases.compute_entropy

MaxEntNucleotideBiases.compute_loglikelihood

MaxEntNucleotideBiases.compute_logz

MaxEntNucleotideBiases.compute_logz_fast

MaxEntNucleotideBiases.compute_minus_energy

MaxEntNucleotideBiases.compute_pressure

MaxEntNucleotideBiases.compute_symmetrized_kl

MaxEntNucleotideBiases.fitmodel

MaxEntNucleotideBiases.fitmodel

MaxEntNucleotideBiases.gauge_mask_variables

MaxEntNucleotideBiases.gauge_zerosum

MaxEntNucleotideBiases.generate_transfer_matrix

MaxEntNucleotideBiases.generate_transfer_matrix_last

MaxEntNucleotideBiases.get_forces_dict

MaxEntNucleotideBiases.readmodel

MaxEntNucleotideBiases.sample_metropolis

MaxEntNucleotideBiases.writemodel

MaxEntNucleotideBiases._compute_freqs — Method

_compute_freqs(seqs::Vector{String}, motifs::Vector{String})

For each motif in motifs, compute the frequency of observed motifs in each sequence in seqs, then divide by the sequence length and take the average of these intensive fractions over the sequences.

MaxEntNucleotideBiases._mcstep! — Method

_mcstep!(L::Int, mlk::Int, curr_seq::Vector{String}, 
              model::NucleotideModel, 
              beta::Float64)

Performs a MonteCarlo step; L is currseq length, mlk is the maximum length of the motifs considered, currseq is the starting sequence, model specifies the model, and beta is the inverse temperature.

MaxEntNucleotideBiases._preprocess_seqs — Method

_preprocess_seqs(seqs_in::Vector{String})

Pre-process the vector of sequences seqs_in, so that sequences only contains letters A, C, G, T. The pre-processed sequence vector is the returned.

MaxEntNucleotideBiases.compute_entropy — Method

compute_entropy(L::Int, model::NucleotideModel; fast::Bool=false)

Compute the entropy of the model with parameters model_pars and having length L. The computation exploit the derivatives of the partition function, that can be approximated to make it faster (using fast=true), although it should be very quick in any case.

MaxEntNucleotideBiases.compute_loglikelihood — Method

compute_loglikelihood(seq::String, model::NucleotideModel; logZ=missing)

Given a sequence seq and the model parameters model_pars, compute the log-likelihood (energy minus log of Z) of this sequence. logZ can be passed directly if pre-computed, otherwise is it computed each time this function is called.

MaxEntNucleotideBiases.compute_logz — Method

compute_logz(model::NucleotideModel, L::Int)

Compute the partition function of a model of length L through the transfer matrix method.

MaxEntNucleotideBiases.compute_logz_fast — Method

compute_logz_fast(model::NucleotideModel, L::Int)

Compute the partition function of a model of length L by taking the real part of the largest eigenvalue of the transfer matrix.

MaxEntNucleotideBiases.compute_minus_energy — Method

compute_minus_energy(seq::String, model::NucleotideModel)

Given a sequence seq and the model parameters model_pars, compute (minus) the energy of this sequence.

MaxEntNucleotideBiases.compute_pressure — Method

compute_pressure(L::Int, model::NucleotideModel; fast::Bool=false)

The pressure on a genome can be quantified as how the genome is different from a random uniform one. Therefore, this function compute the difference in the entropy of a random uniform model and the model described with model_pars, with lenght L, rescaled with the entropy of the uniform model. Setting fast=true allows for a quicker estimation of the entropy of the model (it can be useful for extremely large sequences).

MaxEntNucleotideBiases.compute_symmetrized_kl — Method

compute_symmetrized_kl(L::Int, model1::NucleotideModel, 
                    model2::NucleotideModel; fast::Bool=false)

Compute the symmetrized version of the Kullback-Leibler divergence between the distributions defined by model1_pars and model2_pars. Setting fast=true allows for a quicker estimation of the KL divergence (it can be useful for extremely large sequences).

MaxEntNucleotideBiases.fitmodel — Function

fitmodel(seqs::Union{Vector{String}, String}, Lmotifs::Int, Lmodel::Union{Int,Missing}=missing; 
              pseudocount_param::Float64=0.0, tolerance::Float64=0.01, max_iter::Int=100, 
              verbose::Bool=false, fast::Union{Bool,String}="auto", ZS_gauge::Bool=true)

Fit the model parameters, which are:

only 1-point functions (fields) if Lmotifs==1;

1-point and 2-point functions (fields, 2-mer forces) if Lmotifs==2;

1-point, 2-point and 3-point functions (fields, 2-mer and 3-mer forces) if Lmotifs==3.

Lmodel is the number of nucleotides used for the inference. If seqs is made of sequences of constant length (or if it is String and not a vector), the default option will use the sequences length as Lmodel; otherwise, the default behaviour is to take Lmodel = 5000.

pseudocount_param is the fraction of weigth coming from random uniform sequences in the computation of the motif's frequences from the data.

If fast, the partition function is estimated through the top eigenvalue of the transfer matrix alone (much faster, but slightly less precise, expecially for short sequences). The default value, "auto", automatically uses the fast evaluation for long sequences.

tolerance and max_iter are parameters for the Newton-Raphson algorithm used to solve the system of equations.

ZS_gauge specifies whether the result has to be put in the zero sum gauge before being returned.

MaxEntNucleotideBiases.fitmodel — Function

fitmodel(freqs::Vector{<:Number}, motifs::Vector{String}, L::Int=5000; 
pseudocount_param::Float64=0.0, tolerance::Float64=0.01, max_iter::Int=100, 
verbose::Bool=false, fast::Union{Bool,String}="auto", ZS_gauge::Bool=true)

Fit the model parameters directly from the frequencies freqs of the motifs motifs. The other parameters are the same as for the function that takes as input the sequences directly, and are documented above.

MaxEntNucleotideBiases.gauge_mask_variables — Method

gauge_mask_variables(motifs::Vector{String})

This function returns a Array{Bool} corresponding to motifs that have to be inferred after all motifs containing Ts are fixed to 0 thanks to gauge transformations.

MaxEntNucleotideBiases.gauge_zerosum — Method

gauge_zerosum(model_pars::NucleotideModel)

This function takes as input a dictionary describing the parmeters of a model and changes the gauge into the "zero-sum" gauge.

MaxEntNucleotideBiases.generate_transfer_matrix — Method

generate_transfer_matrix(model::NucleotideModel)

Return the trasfer matrix used for the computation of the partition function. The first index correspond to the left-most nucleotide, the last index to the right-most one.

MaxEntNucleotideBiases.generate_transfer_matrix_last — Method

generate_transfer_matrix_last(model::NucleotideModel)

Return the last transfer matrix used for the computation of the partition function. The first index correspond to the left-most nucleotide, the last index to the right-most one.

MaxEntNucleotideBiases.get_forces_dict — Method

get_forces_dict(model::NucleotideModel)

Return a dictionary of the form motif => force.

MaxEntNucleotideBiases.readmodel — Method

readmodel(fname::AbstractString)

Read file (that should have been written by writemodel) to get a NucleotideModel, which is returned.

MaxEntNucleotideBiases.sample_metropolis — Method

sample_metropolis(L::Int, model::NucleotideModel; 
                         beta::Float64=1.0, Nsamples::Int=1, Nsteps::Int=1, Ntherm::Int=L*10, 
                         startseq::Union{String,Missing}=missing)

Sample Nsample sequences of length L at inverse temperature beta using a model given in model. Ntherm is the number of MonteCarlo steps done before starting collecting sequences, startseq is the starting sequence.

MaxEntNucleotideBiases.writemodel — Method

writemodel(fname::AbstractString, model::NucleotideModel)

Write model to fname.